+++ src/ethereumProofPointEventType.ts
@@ -1,0 +1,10 @@
+/**
+ * Proof Point event type, the type of an {@link EthereumProofPointEvent}.
+ */
+enum EthereumProofPointEventType {
+  Issued,
+  Committed,
+  Revoked,
+}
+
+export { EthereumProofPointEventType };