+++ src/proofPointResolver.ts
@@ -3,10 +3,23 @@ import { ProofPoint } from "./proofPoint";
 import { ProofPointId, ProofPointIdType } from "./proofPointId";
 import { StorageProvider } from "./storage";
 
+/**
+ * Proof point resolver
+ * An object capable of resolving a @ProofPointId to a @ProofPoint
+ */
 export interface ProofPointResolver {
+  /**
+   * Resolve the given @PRoofPointId
+   * @param id The ID of the proof point to resolve
+   * @returns The @ProofPoint corresponding to the given ID.
+   */
   resolve(id: ProofPointId): Promise<ProofPoint>;
 }
 
+/**
+ * General proof point resolver
+ * A @ProofPointResolver capable of handling all types of @ProofPointId
+ */
 export class GeneralProofPointResolver {
   private _ipfsResolver: IpfsProofPointResolver;
   private _webResolver: WebProofPointResolver;
@@ -28,6 +41,10 @@ export class GeneralProofPointResolver {
   }
 }
 
+/**
+ * Ipfs proof point resolver
+ * A @ProofPointResolver capable of handling @ProofPointId of type Ipfs
+ */
 class IpfsProofPointResolver {
   private _storage: StorageProvider;
 
@@ -47,6 +64,10 @@ class IpfsProofPointResolver {
   }
 }
 
+/**
+ * Ipfs proof point resolver
+ * A @ProofPointResolver capable of handling @ProofPointId of type Web
+ */
 class WebProofPointResolver {
   private _httpClient: HttpClient;