+++ test/ProofPointValidator.test.ts
@@ -99,6 +99,7 @@ describe("ProofPointValidator", () => {
       }`,
       "https://example.com/proof-point/1": `
       {
+        "id": "https://example.com/proof-point/1",
         "@context": [
           "https://www.w3.org/2018/credentials/v1",
           "https://provenance.org/ontology/ptf/v2"
@@ -108,11 +109,34 @@ describe("ProofPointValidator", () => {
           "more": [ "data" ],
           "id": "https://provenance.org/subject1"
         },
-        "issuer": "did:web:example.com",
+        "issuer": "example.com",
         "proof": {
           "proofPurpose": "assertionMethod",
-          "type": "https://open.provenance.org/ontology/ptf/v2/ProvenanceProofTypeWeb1",
-          "verificationMethod": "did:web:example.com"
+          "type": "https://open.provenance.org/ontology/ptf/v2/ProvenanceProofType2",
+          "verificationMethod": "example.com"
+        },
+        "type": [
+          "VerifiableCredential",
+          "https://open.provenance.org/ontology/ptf/v2/CertificationCredential"
+        ]
+      }`,
+      "https://example.com/proof-point/wrong-id": `
+      {
+        "id": "https://example2.com/proof-point/1",
+        "@context": [
+          "https://www.w3.org/2018/credentials/v1",
+          "https://provenance.org/ontology/ptf/v2"
+        ],
+        "credentialSubject": {
+          "some": [ "data" ],
+          "more": [ "data" ],
+          "id": "https://provenance.org/subject1"
+        },
+        "issuer": "example.com",
+        "proof": {
+          "proofPurpose": "assertionMethod",
+          "type": "https://open.provenance.org/ontology/ptf/v2/ProvenanceProofType2",
+          "verificationMethod": "example.com"
         },
         "type": [
           "VerifiableCredential",
@@ -122,6 +146,7 @@ describe("ProofPointValidator", () => {
       "https://example.com/proof-point/invalid": "404 not found",
       "https://example.com/proof-point/wrong-proof-type": `
       {
+        "id": "https://example.com/proof-point/wrong-proof-type",
         "@context": [
           "https://www.w3.org/2018/credentials/v1",
           "https://provenance.org/ontology/ptf/v2"
@@ -131,11 +156,11 @@ describe("ProofPointValidator", () => {
           "more": [ "data" ],
           "id": "https://provenance.org/subject1"
         },
-        "issuer": "did:web:example.com",
+        "issuer": "example.com",
         "proof": {
           "proofPurpose": "assertionMethod",
           "type": "https://open.provenance.org/ontology/ptf/v2/ProvenanceProofType1",
-          "verificationMethod": "did:web:example.com"
+          "verificationMethod": "example.com"
         },
         "type": [
           "VerifiableCredential",
@@ -144,6 +169,7 @@ describe("ProofPointValidator", () => {
       }`,
       "https://example.com/proof-point/wrong-issuer": `
       {
+        "id": "https://example.com/proof-point/wrong-issuer",
         "@context": [
           "https://www.w3.org/2018/credentials/v1",
           "https://provenance.org/ontology/ptf/v2"
@@ -153,11 +179,11 @@ describe("ProofPointValidator", () => {
           "more": [ "data" ],
           "id": "https://provenance.org/subject1"
         },
-        "issuer": "did:web:example2.com",
+        "issuer": "example2.com",
         "proof": {
           "proofPurpose": "assertionMethod",
-          "type": "https://open.provenance.org/ontology/ptf/v2/ProvenanceProofTypeWeb1",
-          "verificationMethod": "did:web:example2.com"
+          "type": "https://open.provenance.org/ontology/ptf/v2/ProvenanceProofType2",
+          "verificationMethod": "example2.com"
         },
         "type": [
           "VerifiableCredential",
@@ -166,6 +192,7 @@ describe("ProofPointValidator", () => {
       }`,
       "https://example.com/proof-point/wrong-verification-method": `
       {
+        "id": "https://example.com/proof-point/wrong-verification-method",
         "@context": [
           "https://www.w3.org/2018/credentials/v1",
           "https://provenance.org/ontology/ptf/v2"
@@ -175,11 +202,11 @@ describe("ProofPointValidator", () => {
           "more": [ "data" ],
           "id": "https://provenance.org/subject1"
         },
-        "issuer": "did:web:example.com",
+        "issuer": "example.com",
         "proof": {
           "proofPurpose": "assertionMethod",
-          "type": "https://open.provenance.org/ontology/ptf/v2/ProvenanceProofTypeWeb1",
-          "verificationMethod": "did:web:example2.com"
+          "type": "https://open.provenance.org/ontology/ptf/v2/ProvenanceProofType2",
+          "verificationMethod": "example2.com"
         },
         "type": [
           "VerifiableCredential",
@@ -391,6 +418,9 @@ describe("ProofPointValidator", () => {
     const validity = await subject.validate(id);
     expect(validity.isValid).to.be.false;
     expect(validity.statusCode).to.eq(ProofPointStatus.NotFound);
+    expect(validity.statusMessage).to.eq(
+      "The Proof Point https://example.com/proof-point/invalid could not be resolved."
+    );
   });
 
   it("web validation with wrong proof type is invalid", async () => {
@@ -400,6 +430,9 @@ describe("ProofPointValidator", () => {
     const validity = await subject.validate(id);
     expect(validity.isValid).to.be.false;
     expect(validity.statusCode).to.eq(ProofPointStatus.BadlyFormed);
+    expect(validity.statusMessage).to.eq(
+      "The Proof Point uses an unsupported proof type."
+    );
   });
 
   it("web validation with wrong issuer is invalid", async () => {
@@ -409,6 +442,9 @@ describe("ProofPointValidator", () => {
     const validity = await subject.validate(id);
     expect(validity.isValid).to.be.false;
     expect(validity.statusCode).to.eq(ProofPointStatus.NotFound);
+    expect(validity.statusMessage).to.eq(
+      "The Proof Point cannot be authenticated."
+    );
   });
 
   it("web validation with wrong verification method is invalid", async () => {
@@ -418,5 +454,18 @@ describe("ProofPointValidator", () => {
     const validity = await subject.validate(id);
     expect(validity.isValid).to.be.false;
     expect(validity.statusCode).to.eq(ProofPointStatus.BadlyFormed);
+    expect(validity.statusMessage).to.eq(
+      "The issuer field does not match the proof.verificationMethod field."
+    );
+  });
+
+  it("web validation with wrong id is invalid", async () => {
+    const id = ProofPointId.parse("https://example.com/proof-point/wrong-id");
+    const validity = await subject.validate(id);
+    expect(validity.isValid).to.be.false;
+    expect(validity.statusCode).to.eq(ProofPointStatus.BadlyFormed);
+    expect(validity.statusMessage).to.eq(
+      "The Proof Point id does not match the source URL."
+    );
   });
 });